requirements.txt
authlib==0.11
Click==7.0
Flask==1.0.3
Flask-Admin==1.5.3
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
passlib==1.7.1
pkg-resources==0.0.0
pytest==4.6.3
sendgrid==6.0.5
Werkzeug==0.15.4
WTForms==2.2.1
google-api-python-client
google-auth